add tariff router
added a router for tariff (v1) moved  endpoint into respective folder
